pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "sqlmodel>=0.0.21",
    "alembic>=1.14",
    "python-dotenv>=1.0",
    "taskiq>=0.11",
    "taskiq-redis>=1.0",
    "taskiq-fastapi>=0.3",
    "PyJWT>=2.8",
    "bcrypt>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
